Fix/Github-Actions

+++ make.sh
@@ -24,26 +24,24 @@ function priv_lazbuild
         git submodule update --init --recursive
         git submodule update --recursive --remote
         while read -r; do
-            if ! (lazbuild --verbose-pkgsearch "${REPLY}"); then
-                if ! (lazbuild --add-package "${REPLY}"); then
-                    if ! [[ -f "use/${REPLY}" ]]; then
-                        declare -A VAR=(
-                            [url]="https://packages.lazarus-ide.org/${REPLY}.zip"
-                            [out]=$(mktemp)
-                        )
-                        wget --output-document "${VAR[out]}" "${VAR[url]}" >/dev/null
-                        unzip -o "${VAR[out]}" -d "use/${REPLY}"
-                        rm --verbose "${VAR[out]}"
-                    fi
+            if [[ -n "${REPLY}" ]] &&
+                ! (lazbuild --verbose-pkgsearch "${REPLY}") &&
+                ! (lazbuild --add-package "${REPLY}") &&
+                ! [[ -f "use/${REPLY}" ]]; then
+                    declare -A VAR=(
+                        [url]="https://packages.lazarus-ide.org/${REPLY}.zip"
+                        [out]=$(mktemp)
+                    )
+                    wget --output-document "${VAR[out]}" "${VAR[url]}" >/dev/null
+                    unzip -o "${VAR[out]}" -d "use/${REPLY}"
+                    rm --verbose "${VAR[out]}"
                 fi
-            fi
         done < 'use/components.txt'
         find 'use' -type 'f' -name '*.lpk' -exec lazbuild --add-package-link {} +
     fi
     find 'src' -type 'f' -name '*.lpi' \
         -exec lazbuild --no-write-project --recursive --no-write-project --build-mode=release {} + 1>&2
 )
-
 
 function priv_dpkg
 (
